Fix mhc dtype bug for hyper_connection with compatible kernel (#2100)

+++ src/paddlefleet/transformer/hyper_connection.py
@@ -544,7 +544,7 @@ class HyperConnectionModule(nn.Layer):
             # nn.Linear uses a different BF16 cuBLAS path for this shape and
             # drifts before the first HC BDA.
             x_2d = x.reshape([-1, nC])
-            weight_out_in = weight.t().contiguous()
+            weight_out_in = weight.t().contiguous().astype(x_2d.dtype)
             proj_2d = paddle.matmul(x_2d, weight_out_in, transpose_y=True)
             proj = proj_2d.reshape([*x.shape[:-1], weight.shape[-1]])
         else:
